+++ src/binding/slepc4py/src/SLEPc/NEP.pyx
@@ -1715,6 +1715,32 @@ cdef class NEP(Object):
         CHKERR( NEPNLEIGSGetKSPs(self.nep, &n, &p) )
         return [ref_KSP(p[i]) for i from 0 <= i <n]
 
+    def setNLEIGSPreconditionerMat(self, Mat P=None):
+        """
+        Sets the matrix to be used to build the preconditioner.
+
+        Parameters
+        ----------
+        P: Mat, optional
+           The matrix that will be used in constructing the preconditioner.
+        """
+        cdef PetscMat Pmat = P.mat if P is not None else <PetscMat>NULL
+        CHKERR( NEPNLEIGSSetPreconditionerMat(self.nep, Pmat) )
+
+    def getNLEIGSPreconditionerMat(self):
+        """
+        Gets the matrix previously set by setNLEIGSPreconditionerMat().
+
+        Returns
+        -------
+        P: Mat
+           The matrix that will be used in constructing the preconditioner.
+        """
+        cdef Mat P = Mat()
+        CHKERR( NEPNLEIGSGetPreconditionerMat(self.nep, &P.mat) )
+        PetscINCREF(P.obj)
+        return P
+
     #
 
     def setCISSExtraction(self, extraction):